simplify demo
